+++ src/RcppExports.cpp
@@ -6,6 +6,20 @@
 
 using namespace Rcpp;
 
+// calc_PGtheta_R
+NumericMatrix calc_PGtheta_R(const NumericMatrix& theta_tilda, Eigen::MatrixXd& theta, const int num_doc, const int num_topics);
+RcppExport SEXP _keyATM_calc_PGtheta_R(SEXP theta_tildaSEXP, SEXP thetaSEXP, SEXP num_docSEXP, SEXP num_topicsSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const NumericMatrix& >::type theta_tilda(theta_tildaSEXP);
+    Rcpp::traits::input_parameter< Eigen::MatrixXd& >::type theta(thetaSEXP);
+    Rcpp::traits::input_parameter< const int >::type num_doc(num_docSEXP);
+    Rcpp::traits::input_parameter< const int >::type num_topics(num_topicsSEXP);
+    rcpp_result_gen = Rcpp::wrap(calc_PGtheta_R(theta_tilda, theta, num_doc, num_topics));
+    return rcpp_result_gen;
+END_RCPP
+}
 // make_wsz_cpp
 List make_wsz_cpp(List docs_, List info_, List initialized_);
 RcppExport SEXP _keyATM_make_wsz_cpp(SEXP docs_SEXP, SEXP info_SEXP, SEXP initialized_SEXP) {
@@ -142,6 +156,7 @@ END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
+    {"_keyATM_calc_PGtheta_R", (DL_FUNC) &_keyATM_calc_PGtheta_R, 4},
     {"_keyATM_make_wsz_cpp", (DL_FUNC) &_keyATM_make_wsz_cpp, 3},
     {"_keyATM_keyATMvb_call", (DL_FUNC) &_keyATM_keyATMvb_call, 1},
     {"_keyATM_read_dfm_cpp", (DL_FUNC) &_keyATM_read_dfm_cpp, 4},